custom hook refactor (continued)
