+++ src/store/userProgress/reducer.js
@@ -1,28 +1,29 @@
 import Immutable from 'seamless-immutable';
 import {
-    SET_USER_DATA,
-    FETCH_USER_DATA,
-    FETCH_USER_DATA_LOADING,
-    FETCH_USER_DATA_DONE,
-    INIT_USER_DATA,
-    FETCH_USER_DATA_REJECTED,
-    INIT_USER_DATA_REJECTED,
-    SET_STEP_CALCULATE_2,
-    UPDATE_USER_DATA,
-    UPDATE_USER_DATA_LOADING,
-    UPDATE_USER_DATA_DONE,
-    UPDATE_USER_DATA_REJECTED,
-    SOLAR_CALCULATIONS,
-    SOLAR_CALCULATIONS_LOADING,
-    SOLAR_CALCULATIONS_DONE,
-    SOLAR_CALCULATIONS_REJECTED,
-    LOADING,
-    CALCULATE1,
-    CALCULATE2,
-    CALCULATE4,
-    INITIAL_AVG_BILL
+  SET_USER_DATA,
+  FETCH_USER_DATA,
+  FETCH_USER_DATA_LOADING,
+  FETCH_USER_DATA_DONE,
+  INIT_USER_DATA,
+  FETCH_USER_DATA_REJECTED,
+  INIT_USER_DATA_REJECTED,
+  SET_STEP_CALCULATE_2,
+  UPDATE_USER_DATA,
+  UPDATE_USER_DATA_LOADING,
+  UPDATE_USER_DATA_DONE,
+  UPDATE_USER_DATA_REJECTED,
+  SOLAR_CALCULATIONS,
+  SOLAR_CALCULATIONS_LOADING,
+  SOLAR_CALCULATIONS_DONE,
+  SOLAR_CALCULATIONS_REJECTED,
+  LOADING,
+  CALCULATE1,
+  CALCULATE2,
+  CALCULATE3,
+  CALCULATE4,
+  RESULT1,
+  INITIAL_AVG_BILL,
 } from './actions';
-import { nextStepLookup } from '../../containers/Calculate/lookupObjects';
 
 export const initialState = Immutable({
   loading: true,
@@ -35,6 +36,11 @@ export const initialState = Immutable({
   },
   error: '',
 });
+
+export const getNextStep = (userData) => {
+  if(userData.step === CALCULATE1) return CALCULATE1;
+  return userData.savings ? RESULT1 : CALCULATE3;
+}
 
 export default function reduce(state = initialState, action = {}) {
   switch (action.type) {
@@ -62,7 +68,7 @@ export default function reduce(state = initialState, action = {}) {
         : {};
       return state.merge({ updateLoading: false, ...reloadState });
     case UPDATE_USER_DATA:
-      return state.merge({ userData: { ...state.userData, ...action.payload, step: nextStepLookup(action.payload) } });
+      return state.merge({ userData: { ...state.userData, ...action.payload, step: getNextStep(action.payload) } });
     case UPDATE_USER_DATA_REJECTED:
       return state.merge({ error: action.payload });
     case SOLAR_CALCULATIONS_LOADING:
